# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indonesian Stock Analysis API",
    "description": "Production-grade realtime and historical analysis API for Indonesian (IDX) equities.\n\nAll `/api/v1` data endpoints require authentication (`X-API-Key` header or `Authorization: Bearer <jwt>`) **and** an allow-listed browser origin. Responses use a uniform `{meta, data}` envelope; errors use a uniform `{error}` envelope.",
    "contact": {
      "name": "teguh-prasetyo.com",
      "url": "https://www.teguh-prasetyo.com/"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.teguh-prasetyo.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/stock/{symbol}": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Realtime quote for an IDX symbol",
        "operationId": "get_quote_api_v1_stock__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 12,
              "description": "IDX ticker symbol.",
              "examples": [
                "BBCA"
              ],
              "title": "Symbol"
            },
            "description": "IDX ticker symbol."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Quote_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{symbol}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Historical OHLCV with pagination",
        "operationId": "get_history_api_v1_history__symbol__get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 12,
              "description": "IDX ticker symbol.",
              "examples": [
                "BBCA"
              ],
              "title": "Symbol"
            },
            "description": "IDX ticker symbol."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (inclusive), Jakarta time.",
              "title": "Start"
            },
            "description": "Start date (inclusive), Jakarta time."
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (inclusive), Jakarta time.",
              "title": "End"
            },
            "description": "End date (inclusive), Jakarta time."
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Interval",
              "description": "Sampling interval.",
              "default": "1d"
            },
            "description": "Sampling interval."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "description": "1-based page index.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page index."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Rows per page.",
              "default": 200,
              "title": "Page Size"
            },
            "description": "Rows per page."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_HistoryData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_HistoryData_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "$ref": "#/components/schemas/HistoryData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[HistoryData]"
      },
...
      "Envelope_Quote_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "data": {
            "$ref": "#/components/schemas/Quote"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta",
          "data"
        ],
        "title": "Envelope[Quote]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryData": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "bars": {
            "items": {
              "$ref": "#/components/schemas/OHLCVBar"
            },
            "type": "array",
            "title": "Bars"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "interval",
          "start",
          "end",
          "pagination",
          "bars"
        ],
        "title": "HistoryData",
        "description": "Paginated historical OHLCV series."
      },
      "Interval": {
        "type": "string",
        "enum": [
          "1m",
          "5m",
          "15m",
          "30m",
          "1h",
          "1d",
          "1wk",
          "1mo"
        ],
        "title": "Interval",
        "description": "Supported sampling intervals (mapped to provider-native values)."
      },
...
      "OHLCVBar": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "timestamp",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "OHLCVBar",
        "description": "A single open/high/low/close/volume observation."
      },
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta"
      },
      "Quote": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "price": {
            "type": "number",
            "title": "Price",
            "description": "Most recent traded/last price."
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change",
            "description": "Absolute change vs previous close."
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent",
            "description": "Percentage change vs previous close."
          },
          "day_open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Open"
          },
          "day_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day High"
          },
          "day_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Low"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "market_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Time",
            "description": "Exchange timestamp of the quote."
          },
          "is_market_open": {
            "type": "boolean",
            "title": "Is Market Open",
            "description": "Indicative IDX session-open flag."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "price",
          "is_market_open"
        ],
        "title": "Quote",
        "description": "A realtime (or last-known) quote snapshot."
      },
...
      "ResponseMeta": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol",
            "description": "Canonical IDX symbol, e.g. BBCA.JK."
          },
          "exchange": {
            "type": "string",
            "title": "Exchange",
            "description": "Source exchange identifier.",
            "default": "IDX"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency of all monetary values.",
            "default": "IDR"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "description": "Time zone of all timestamps.",
            "default": "Asia/Jakarta"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Upstream data provider that served the data."
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "UTC instant the response was produced."
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "Whether the payload was served from cache.",
            "default": false
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Data is provided for informational purposes only and may be delayed. It is not investment advice."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "generated_at"
        ],
        "title": "ResponseMeta",
        "description": "Provenance metadata attached to every successful data response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}